+++ lib/index.jsx
@@ -3,6 +3,7 @@ import { useState } from 'react';
 import TabContent from './TabContent';
 import { InjectorContext } from './Context/InjectorContext';
 import { FilterContext } from './Context/FilterContext';
+import { ScopeExpandProvider } from './Context/ScopeExpandContext';
 
 /**
  *
@@ -18,13 +19,16 @@ export default function(props) {
 
   const search = useState({
     search: '',
-    filterType: 'all'
+    filterType: 'all',
+    selectedElements: []
   });
 
   return (
     <InjectorContext.Provider value={ injector }>
       <FilterContext.Provider value={ search }>
-        <TabContent />
+        <ScopeExpandProvider>
+          <TabContent />
+        </ScopeExpandProvider>
       </FilterContext.Provider>
     </InjectorContext.Provider>
   );