+++ lib/Components/FilterBar/FilterBar.jsx
@@ -1,5 +1,6 @@
 import ElementFilter from './ElementFilter';
 import WrittenOnlyToggle from './WrittenOnlyToggle';
+import CollapseAllButton from './CollapseAllButton';
 
 import './FilterBar.scss';
 
@@ -9,6 +10,7 @@ export default function FilterBar({ variables }) {
       <ElementFilter variables={ variables } />
       <div className="bio-vo-filter-bar-actions">
         <WrittenOnlyToggle />
+        <CollapseAllButton />
       </div>
     </div>
   );