+++ lib/TabContent.jsx
@@ -6,10 +6,11 @@ import EmptySearch from './Components/EmptySearch';
 
 import { useVariables } from './hooks/useVariables';
 import Search from './Components/Search';
+import useCanvasSync from './hooks/useCanvasSync';
 
 export default function(props) {
-
   const { rawVariables, availableVariables } = useVariables();
+  useCanvasSync();
 
   return <div className="bio-vo-tab-content">
     <Search />