+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to the [@bpmn-io/variable-outline](https://github.com/bpmn-i
 * `FEAT`: make layout vertical for variables ([#16](https://github.com/bpmn-io/variable-outline/pull/16))
 * `FEAT`: restyle search bar ([#18](https://github.com/bpmn-io/variable-outline/pull/18))
 * `FEAT`: add elements filter ([#19](https://github.com/bpmn-io/variable-outline/pull/19))
+* `FEAT`: add bi-directional sync between canvas selection and elements filters ([#20](https://github.com/bpmn-io/variable-outline/pull/20))
 
 ## 2.0.1